Features/i ingredient data

+++ JP_PantryAid/PantryAid.Core/Models/IngredientItem.cs
@@ -7,10 +7,10 @@ namespace PantryAid.Core.Models
     public class IngredientItem
     {
 		public Ingredient Ingredient { get; set; }
-		public float Quantity { get; set; }
+		public double Quantity { get; set; }
 		public Measurements MeasurementType { get; set; }
 
-		public IngredientItem(Ingredient ingr, float amount, Measurements measure)
+		public IngredientItem(Ingredient ingr, double amount, Measurements measure)
 		{
 			Ingredient = ingr;
 			Quantity = amount;
